fix: 인증과 사용자 기능 분리
