+++ src/main/java/com/spring/study/springsecurity/user/service/AuthService.java
@@ -1,0 +1,30 @@
+package com.spring.study.springsecurity.user.service;
+
+import com.spring.study.springsecurity.exception.CustomException;
+import com.spring.study.springsecurity.exception.ErrorCode;
+import com.spring.study.springsecurity.user.dto.SignupRequestDto;
+import com.spring.study.springsecurity.user.dto.SignupResponseDto;
+import com.spring.study.springsecurity.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@RequiredArgsConstructor
+public class AuthService {
+
+  private final UserRepository userRepository;
+  private final PasswordEncoder passwordEncoder;
+
+  @Transactional
+  public SignupResponseDto signupUser(SignupRequestDto request) {
+
+    if (userRepository.existsByEmail(request.getEmail())) {
+      throw new CustomException(ErrorCode.DUPLICATE_EMAIL);
+    }
+
+    return SignupResponseDto.from(
+        userRepository.save(request.toEntity(passwordEncoder)));
+  }
+}
